requirements.txt
torch>=1.2
torchvision
numpy
pillow
yaml
requests